% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{england_ltla_shape}
\alias{england_ltla_shape}
\title{LTLA Shapefile for England}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 317 rows and 11 columns.
}
\usage{
england_ltla_shape
}
\description{
LTLA Shapefile for England
}
\keyword{datasets}
