% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{england_utla_shape}
\alias{england_utla_shape}
\title{UTLA Shapefile for England}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 151 rows and 11 columns.
}
\usage{
england_utla_shape
}
\description{
UTLA Shapefile for England
}
\keyword{datasets}
