% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Link Mapping to Trust and UTLA Names}
\usage{
get_names(mapping)
}
\arguments{
\item{mapping}{data.frame containing trust_code, p_trust, utla_code and p_utla.
Defaults to \code{trust_utla_mapping} if not supplied.}
}
\value{
A data.frame containing a UTLA to trust level admissions map combined
meaningful names.
}
\description{
Link Mapping to Trust and UTLA Names
}
