% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mappings.R
\name{load_mapping}
\alias{load_mapping}
\title{Load local authority to Trust mappings}
\usage{
load_mapping(scale, source)
}
\arguments{
\item{scale}{Character string defining the geographical scale. Supported options are "ltla" (lower-tier local authority) and "utla" (upper-tier local authority).}

\item{source}{Character string defining the souce of the mapping. supported options are "link" (linked COVID-19 cases and admissions) and "sus" (Secondary Uses Service, originally the only mapping available).}
}
\value{
A data.frame with the following columns: \code{geo_code} (9-digit LTLA or UTLA ID); \code{trust_code}; \code{n} (the number of reported admissions); \code{p_geo} (the proportion of all admissions from a LTLA/UTLA that go to a given Trust); \code{p_trust} (the proportion of all admissions to a Trust that come from a given LTLA/UTLA); \code{source} ("SUS" or "Link"); and \code{level} ("ltla" or "utla").
}
\description{
Load LTLA- or UTLA-Trust mappings from two data sources: HES (Hospital Episodes Statistics) until September 2020, and linked COVID-19 cases and admissions June 2020 - May 2021. Replaces previous datasets trust_ltla_mapping and trust_utla_mapping (the HES mappings).
}
