% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ltla_names}
\alias{ltla_names}
\title{Linking File Between LTLA Codes and Names}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 340 rows and 2 columns.
}
\usage{
ltla_names
}
\description{
Linking File Between LTLA Codes and Names
}
\keyword{datasets}
