% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mapping.R
\name{summarise_mapping}
\alias{summarise_mapping}
\title{Summarise Trust Mapping}
\usage{
summarise_mapping(trust = NULL, utla = NULL, mapping, shapefile)
}
\arguments{
\item{trust}{A character string indicating a trust of interest.}

\item{utla}{A character string indicating the UTLA of interest. Only used if
\code{trust} is not specified.}

\item{mapping}{data.frame containing trust_code, p_trust, utla_code and p_utla.
Defaults to \code{trust_utla_mapping} if not supplied.}

\item{shapefile}{A shapefile defaults to \code{uk_utla_shape} if not supplied.}
}
\value{
A table and optional map summarising the admissions mapping.
}
\description{
Summarise the Trust mapping (currently only supports the UTLA mapping) both
graphically and in a table.
}
