% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{trust_ltla_mapping}
\alias{trust_ltla_mapping}
\title{Mapping Between NHS Trust and LTLA}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 590 rows and 4 columns.
}
\usage{
trust_ltla_mapping
}
\description{
Mapping Between NHS Trust and LTLA
}
\keyword{datasets}
