% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{trust_mergers}
\alias{trust_mergers}
\title{List of Trust Mergers}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 11 rows and 3 columns.
}
\usage{
trust_mergers
}
\description{
List of Trust Mergers
}
\keyword{datasets}
