% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{trust_names}
\alias{trust_names}
\title{Linking File Between NHS Trust Codes and Names}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 228 rows and 2 columns.
}
\usage{
trust_names
}
\description{
Linking File Between NHS Trust Codes and Names
}
\keyword{datasets}
