% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{trust_utla_mapping}
\alias{trust_utla_mapping}
\title{Mapping Between NHS Trust and UTLA}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 424 rows and 4 columns.
}
\usage{
trust_utla_mapping
}
\description{
Mapping Between NHS Trust and UTLA
}
\keyword{datasets}
