% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admissions.R
\name{get_admissions}
\alias{get_admissions}
\title{Get Hospital Admissions}
\usage{
get_admissions(
  keep_vars = "new_adm",
  level = "trust",
  release_date = Sys.Date(),
  mapping,
  geo_names
)
}
\arguments{
\item{keep_vars}{Character string, defaulting to "new_adm" (first-time COVID-19 hospital admissions). Defines which variables to keep from the raw data. Other supported options are: "all_adm" (for all COVID-19 hospital admissions), and "all_bed" (for all COVID-19 beds occupied). Multiple values allowed.}

\item{level}{Character string, defaulting to "trust". Defines the level of aggregation
at which to return the data. Other supported options are "utla" for UTLA level admissions
or "ltla" for LTLA level admissions.}

\item{release_date}{Date, release date of data to download. Will automatically find
the Thursday prior to the date specified.}

\item{mapping}{A data.frame containing geo_code, trust_code, p_geo and p_trust.}

\item{geo_names}{A data.frame containing \code{geo_code} and \code{geo_name}. Used to
assign meaningful to geographies.}
}
\value{
A data.frame of daily admissions and/or bed occupancy data, reported at the Trust, LTLA or UTLA level. Note that new admissions ("new_adm") are called "admissions" in the data.frame to be consistent with a previous version of this function.
}
\description{
Downloads hospital admissions by Hospital trust using
\code{download_trust_data} and then optionally aggregates to either LTLA or UTLA
level. This can be done either with the built in mapping or a user supplied mapping.
}
