% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_admissions.R
\name{map_admissions}
\alias{map_admissions}
\title{Map Admissions}
\usage{
map_admissions(admissions, shapefile, date, scale_fill)
}
\arguments{
\item{admissions}{A data.frame of admissions data as produced by \code{get_dadmissions}.
Must contain the following variables: \code{geo_code}, \code{date}, and \code{admissions}}

\item{shapefile}{A shapefile defaults to \code{england_utla_shape} if not supplied.}

\item{date}{A date variable (or vector) indicating when to plot data for.}

\item{scale_fill}{A \code{ggplot2} \code{scale_fill} used to define the fill colours
used in the map. The default is:
\code{scale_fill_viridis_c(option = "viridis", direction = -1, na.value = "grey80")}.}
}
\value{
A map of Covid-19 admissions in England
}
\description{
Map NHS Hospital aggregated admissions. By default this maps the
latest data provided.
}
